📝Refactored to Bluetooth class
